Separate files for stencil type operations to restore compilation with llvm24.0.0
